Implemented SemiEvalDataIO._load_data
